move getActivePath to separate file
